refactor: move the kv related logic from the server.go to separate file and separate unit tests
